refactor: structuring component into folders with css file- #31
