+++ src/App.tsx
@@ -1,11 +1,12 @@
 import { NavLink } from "react-router";
 import { Tree } from "./components/Tree.tsx";
-import Layout from "./components/Layout";
+
 import { TreeContext, createTreeManagerFromModelNodes, getLayout } from "./contexts/TreeContext.ts"
 import { ModelDefinition, getModelNodes, applyLayers } from "./modules/LayeredModel.ts"
 import { baseModel } from "../examples/example-models.ts"
 import "@xyflow/react/dist/style.css";
 import "./App.css";
+import Layout from "./components/Layout/Layout";
 
 const App = () => {
   const modelDefintion: ModelDefinition = applyLayers(baseModel, []);
