+++ src/components/Layout/Layout.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import './Layout.css';
 
 
 const Layout: React.FC<{ children: React.ReactNode }> = ({ children }) => {
