+++ src/index.css
@@ -1,23 +1,3 @@
-.layout {
-    display: flex;
-    flex-direction: column;
-    min-height: 100vh;
-}
-
-header, footer {
-    background-color: #f8f8f8;
-    padding: 1rem;
-    text-align: center;
-}
-
-main {
-    flex: 1;
-    padding: 1rem;
-    display: flex;
-    justify-content: center;
-    align-items: stretch;
-    overflow: hidden;
-}
 .tree-container {
     width: 100%;
     flex: 1;
